requirements.txt
azure-identity
azure-ai-projects
azure-search-documents

openai

autogen
autogen-core
autogen-agentchat


python-dotenv
chainlit